pyproject.toml
[project]
name = "mutenix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
     "hidapi",
     "websockets",
     "pydantic",
     "requests",
     "pywin32; platform_system=='Windows'",
     "pywinauto; platform_system=='Windows'",
     "aiohttp",
     "tqdm",
     "semver",
     "types-requests>=2.32.0.20241016",
     "pytest-asyncio>=0.25.0",
]

[project.scripts]
hello = "mutenix:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest-cov>=6.0.0",
    "pytest-timeout>=2.3.1",
    "pytest>=8.3.4",
]
